+++ src/lib.rs
@@ -1,1 +1,2 @@
-
+// Export the media_file_info module
+pub mod media_file_info;